+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Unreleased
 
+# 2.12.0 (2019-08-26)
+
 * Allow `Person` to _optionally_ accept a `caption` prop
 * Add `ThickBanner` component